Moved the block codes to two files and modified the for loop
