+++ scripts.js
@@ -1,890 +1,19 @@
-import { categoryColors, blockCategory } from "./scripts/blockConfiguration.js";
-import { getBlockDropdownList, getBlockProperties, getCategoryByBlockID, createBlockLabel } from "./scripts/blockProperties.js";
 import { EditorView, basicSetup } from "codemirror";
 import { python } from "@codemirror/lang-python";
 
-let blockCounter = 0;
+import { blockCategory } from "./scripts/blockConfiguration.js";
+import { getBlockProperties } from "./scripts/blockProperties.js";
+import { 
+  createCategoryButtons, 
+  newBlock,
+  clearDropHighlights
+} from "./scripts/blockCreation.js";
+
 let dragged = null;
 let highlightedBlock = null;
-let isPythonView = false;
-let userVariables = [];
 
-// ==========================
-// 3. Block Management Functions
-// ==========================
-
-// Function to dynamically create buttons and assign background colors to categories
-function createCategoryButtons(blockCategory) {
-  for (const [categoryName, categoryData] of Object.entries(blockCategory)) {
-    const categoryContainer = document.getElementById(categoryName);
-
-    // Check if the category container exists
-    if (!categoryContainer) {
-      console.warn(`No container found for category: ${categoryName}`);
-      continue;
-    }
-
-    // Apply the background color to the category header
-    const categoryHeader =
-      categoryContainer.parentElement.querySelector(".category-header");
-    const color = categoryColors[categoryName] || "#cccccc";
-
-    // Set default color for Variable Declaration block
-    if (categoryName === "Variable Declaration") {
-      categoryHeader.style.backgroundColor = "#cccccc";
-    } else {
-      categoryHeader.style.backgroundColor = color;
-    }
-
-    // Add the onclick event listener to the category header
-    categoryHeader.addEventListener("click", function () {
-      toggleCategory(categoryName);
-    });
-
-    // Iterate through each element in the category
-    categoryData.elements.forEach((element) => {
-      // Skip creating Variable Operations and Variable Blocks if no variables exist
-      if (
-        (element.blockID === "varOps" || element.blockID === "variableBlock") &&
-        userVariables.length === 0
-      ) {
-        return; // Skip this element
-      }
-
-      const button = document.createElement("button");
-      button.id = `${element.blockID}Button`;
-      button.name = element.name;
-      button.innerText = `${element.name}`;
-
-      // Apply category color to the button
-      // Set default color for Variable Declaration block buttons
-      if (element.blockID === "varDeclOps") {
-        button.style.backgroundColor = "#cccccc";
-      } else {
-        button.style.backgroundColor = color;
-      }
-
-      // Add the description to the button's title attribute
-      button.title = element.description; // Tooltip text showing description
-
-      // Add an onclick handler to call newBlock
-      button.onclick = function () {
-        // Call the newBlock function with the appropriate parameters
-        newBlock(element.blockID);
-      };
-
-      // Append the button to the category container
-      categoryContainer.appendChild(button);
-    });
-  }
-}
 // Call the function to create the buttons
 createCategoryButtons(blockCategory);
-
-function refreshCategoryButtons() {
-  // Clear all category containers
-  for (const categoryName of Object.keys(blockCategory)) {
-    const categoryContainer = document.getElementById(categoryName);
-    if (categoryContainer) {
-      categoryContainer.innerHTML = ""; // Clear the container
-    }
-  }
-
-  // Recreate the buttons
-  createCategoryButtons(blockCategory);
-}
-
-function newBlock(blockID) {
-  if (isPythonView) {
-    console.warn("Cannot create a new block in Python view.");
-    return; // Exit the function early
-  }
-
-  const container = document.getElementById("box-container");
-  const newBlock = document.createElement("div");
-  newBlock.classList.add("box");
-  newBlock.id = "box_" + ++blockCounter; // Increment block counter and set ID
-  newBlock.dataset.blockID = blockID; // Store the blockID in the dataset
-
-  // Calculate and set the depth
-  const parentDepth = Number(container.getAttribute("data-blockDepth")) || 0;
-  newBlock.dataset.blockDepth = parentDepth + 1; // Parent depth + 1 for the new block
-
-  // Set block properties
-  const { blockCategoryColor, childElement } = getBlockProperties(blockID);
-  newBlock.style.backgroundColor = blockCategoryColor; // Apply the category color
-
-  createBlockLabel(newBlock, blockID);
-
-  // Handle different block types
-  if (blockID === "mathBlock" || blockID === "comparisonBlock" || blockID === "varOps") {
-    handleMathOrComparisonOrVariableBlock(newBlock, blockID);
-  } else if (["if", "while", "for"].includes(blockID)) {
-    handleControlBlock(newBlock, blockID);
-  } else if (blockID === "mathText") {
-    createInputBlock(newBlock, "0", "math-input", "blockValue", blockID);
-  } else if (blockID === "printText") {
-    createInputBlock(newBlock, "Enter Text", "text-input", "blockValue", blockID);
-  } else if (blockID === "varDeclOps") {
-    handleVariableDeclarationBlock(newBlock);
-  } else if (blockID === "variableBlock") {
-    handleVariableBlock(newBlock);
-  } else if (
-    blockID === "arithmeticOps" ||
-    blockID === "comparisonOps" ||
-    blockID === "logicalOps"
-  ) {
-    handleOperatorBlock(newBlock, blockID);
-  } else if(blockID === "print") {
-    handlePrintBlock(newBlock, blockID);
-  } else if (blockID === "movement") {
-    handleMovementBlock(newBlock, blockID);
-  } else if (
-    blockID === "mathConstants" ||
-    blockID === "roundingTruncation" ||
-    blockID === "absSign" ||
-    blockID === "numberTheory" ||
-    blockID === "sumProd" ||
-    blockID === "floatManipulation" ||
-    blockID === "comparisonValidation" ||
-    blockID === "remainderDivision" ||
-    blockID === "logExpFunctions" ||
-    blockID === "trigFunctions" ||
-    blockID === "hyperbolicFunctions" ||
-    blockID === "specialFunctions"
-    
-  ) {
-    handleMathFunctionBlock(newBlock, blockID); // Handle math function blocks
-  } else {
-    handleDefaultBlock(newBlock, blockID);
-  }
-
-  if (blockID !== "varDeclOps") {
-    appendChildElement(newBlock, childElement);
-    container.appendChild(newBlock);
-    addBlockInteractivity(newBlock);
-    updateLineNumbers();
-  }
-  return newBlock.id;
-}
-
-// Function to remove a block by its ID
-function removeBlock(blockId) {
-  const block = document.getElementById(blockId);
-  if (block) {
-    block.remove(); // Remove the block from the DOM
-    updateLineNumbers(); // Update line numbers after removal
-  }
-}
-
-// Function to update the line numbers based on the number of blocks
-function updateLineNumbers() {
-  const codeLinesContainer = document.querySelector(".code-lines");
-  const blocks = document.querySelectorAll("#box-container .box");
-
-  // Clear existing line numbers
-  codeLinesContainer.innerHTML = "";
-
-  // Create new line numbers based on the number of blocks
-  const totalLines = blocks.length + 1; // +1 for the extra empty line at the bottom
-  for (let i = 1; i <= totalLines; i++) {
-    const lineNumber = document.createElement("div");
-    lineNumber.classList.add("code-line");
-    if (i === totalLines) {
-      // This is the extra empty line at the bottom
-      lineNumber.textContent = i;
-    } else {
-      lineNumber.textContent = i; // Line numbers start from 1
-    }
-    codeLinesContainer.appendChild(lineNumber);
-  }
-}
-
-// ==========================
-// 4. Handling Different Block Types
-// ==========================
-
-function handleDefaultBlock(block, blockID) {
-  const blockName = blockCategory[getCategoryByBlockID(blockID)].elements.find(
-    (element) => element.blockID === blockID
-  ).name;
-
-  const blockLabel = document.createElement("span");
-  blockLabel.textContent = `${blockName}`;
-  block.appendChild(blockLabel);
-}
-
-function handlePrintBlock(newBlock, blockID){
-  const printLabel = document.createElement("span");
-  printLabel.textContent = "print";
-  newBlock.appendChild(printLabel);
-
-  const childContainer = createChildBoxHorizontal(newBlock.id, blockID);
-  newBlock.appendChild(childContainer);
-}
-
-function handleMovementBlock(block, blockID) {
-  const turtlePrefix = document.createElement("span"); //temp text
-  turtlePrefix.textContent = "turtle.";
-  block.appendChild(turtlePrefix);
-
-  const dropdown = createOperatorDropdown(blockID);
-  block.appendChild(dropdown);
-
-  if (blockID != "movement") {
-    const childContainer = createChildBoxHorizontal(block.id, blockID);
-    block.appendChild(childContainer);
-  }
-
-  // Add a horizontal child block for the value input
-  const childContainer = createChildBoxHorizontal(block.id, blockID);
-  block.appendChild(childContainer);
-}
-
-function handleOperatorBlock(block, blockID) {
-  const blockName = blockCategory[getCategoryByBlockID(blockID)].elements.find(
-    (element) => element.blockID === blockID
-  ).name;
-
-  // Create and append the dropdown list
-  const dropdown = createOperatorDropdown(blockID);
-  block.appendChild(dropdown);
-}
-
-function handleMathFunctionBlock(block, blockID) {
-  const mathPrefix = document.createElement("span");
-  mathPrefix.textContent = "math.";
-  block.appendChild(mathPrefix);
-
-  const dropdown = createOperatorDropdown(blockID);
-  block.appendChild(dropdown);
-  if (blockID != "mathConstants") {
-    // Add a horizontal child block for the value input
-    const childContainer = createChildBoxHorizontal(block.id, blockID);
-    block.appendChild(childContainer);
-  }
-
-}
-
-function handleMathOrComparisonOrVariableBlock(block, blockID) {
-  const container = document.createElement("div");
-  container.classList.add("childBox-Container-Horizontal");
-
-  // Create the left-hand side container for the variable or math/comparison block
-  const leftContainer = createChildBoxHorizontal(block.id, blockID);
-  container.appendChild(leftContainer);
-
-  // Create and append the operator dropdown
-  const operatorDropdown = createOperatorDropdown(blockID);
-  operatorDropdown.dataset.dropdownType = "operator";
-  container.appendChild(operatorDropdown);
-
-  // Create the right-hand side container for the value block
-  const rightContainer = createChildBoxHorizontal(block.id, blockID);
-  container.appendChild(rightContainer);
-
-  block.appendChild(container);
-
-  // Update data attributes based on changes in the operator dropdown
-  operatorDropdown.addEventListener("change", function () {
-    block.dataset.blockOperator = operatorDropdown.value;
-  });
-
-  // Handle changes in the left-hand side container (variable or math/comparison block)
-  leftContainer.addEventListener("change", function () {
-    const leftBlock = leftContainer.querySelector(".box");
-    if (leftBlock) {
-      block.dataset.block1Value = leftBlock.dataset.blockValue || "";
-    }
-  });
-
-  // Handle changes in the right-hand side container (value block)
-  rightContainer.addEventListener("change", function () {
-    const rightBlock = rightContainer.querySelector(".box");
-    if (rightBlock) {
-      block.dataset.block2Value = rightBlock.dataset.blockValue || "";
-    }
-  });
-}
-
-function handleControlBlock(block, blockID) {
-  if (blockID === "if") {
-    handleIfBlock(block, blockID);
-  } else if (blockID === "for" || blockID === "while") {
-    handleLoopBlocks(block, blockID);
-  }
-}
-
-function handleIfBlock(block, blockID) {
-  const topChildBox = createChildBoxHorizontal(block.id, blockID);
-  const topLabel = document.createElement("span");
-  topLabel.classList.add("block-top-label");
-  topLabel.textContent = `${blockID.charAt(0).toUpperCase() + blockID.slice(1)}:`;
-  block.appendChild(topLabel);
-  block.appendChild(topChildBox);
-  addBlockInteractivity(block);
-  updateLineNumbers();
-}
-
-function handleLoopBlocks(block, blockID) {
-  const topChildBox = createChildBoxHorizontal(block.id, blockID);
-  const topLabel = document.createElement("span");
-  topLabel.classList.add("block-top-label");
-  topLabel.textContent = `${blockID.charAt(0).toUpperCase() + blockID.slice(1)}:`;
-  block.appendChild(topLabel);
-  block.appendChild(topChildBox);
-
-  if (blockID === "for") {
-    const extraTopChildBox = createChildBoxHorizontal(block.id, blockID);
-    const extraTopLabel = document.createElement("span");
-    extraTopLabel.classList.add("block-top-label");
-    extraTopLabel.textContent = "Range:";
-    block.appendChild(extraTopLabel);
-    block.appendChild(extraTopChildBox);
-  }
-}
-
-function handleVariableDeclarationBlock(block) {
-  const variableName = prompt("Enter a new variable name:");
-  if (!variableName) return;
-
-  if (!userVariables.includes(variableName)) {
-    userVariables.push(variableName);
-    updateUserVariableDropdowns(); // Update dropdowns
-    refreshCategoryButtons(); // Refresh the category buttons
-  }
-}
-
-function handleVariableBlock(block) {
-  const container = document.createElement("div");
-  container.classList.add("childBox-Container-Horizontal");
-
-  // Variable Dropdown
-  const variableDropdown = document.createElement("select");
-  variableDropdown.classList.add("block-dropdown");
-  variableDropdown.setAttribute("data-type", "variable"); // Add this line
-  userVariables.forEach((varName) => {
-    const option = document.createElement("option");
-    option.value = varName;
-    option.textContent = varName;
-    variableDropdown.appendChild(option);
-  });
-
-  variableDropdown.value = "---"; // Default value
-  variableDropdown.addEventListener("change", function () {
-    block.dataset.blockValue = variableDropdown.value;
-  });
-
-  container.appendChild(variableDropdown);
-  block.appendChild(container);
-}
-
-function handleElseIfOption(block, elifElseDiv, dropdown, plusIcon) {
-  if (elifElseDiv.children.length === 1 && elifElseDiv.contains(plusIcon)) {
-    elifElseDiv.remove(); // Remove the plus icon container if it exists
-  }
-
-  // Create the new "else if" block
-  const newElifElseDiv = createElifElseDiv("elif");
-  const elseIfSpan = createSpan("else if:");
-  newElifElseDiv.appendChild(elseIfSpan);
-
-  const horizontalChildBox = createChildBoxHorizontal(block.id, block.dataset.blockID);
-  newElifElseDiv.appendChild(horizontalChildBox);
-
-  // Append the new "else if" block to the parent block
-  block.appendChild(newElifElseDiv);
-  appendChildElement(block, "block");
-
-  // Move the "else" block and plus icon to the end
-  const elseBlock = block.querySelector('.elif-else[data-elif-else-type="else"]');
-  const plusIconDiv = block.querySelector('.elif-else[data-elif-else-type="plus"]');
-
-  if (elseBlock) {
-    block.appendChild(elseBlock); // Move the "else" block to the end
-  }
-  if (plusIconDiv) {
-    block.appendChild(plusIconDiv); // Move the plus icon to the end
-  } else {
-    // If the plus icon doesn't exist, create it and append it to the end
-    const newPlusIcon = createPlusIcon();
-    const newElifElseDivForPlus = createElifElseDiv("plus");
-    newElifElseDivForPlus.appendChild(newPlusIcon);
-    block.appendChild(newElifElseDivForPlus);
-
-    // Set up the dropdown menu for the new plus icon
-    setupDropdownMenu(newPlusIcon, block, newElifElseDivForPlus);
-  }
-
-  // Reset and update the IDs for all elif-else blocks
-  resetAndUpdateElifElseIds(block);
-
-  // Increment the "else if" counter
-  block.dataset.elseIfCount = parseInt(block.dataset.elseIfCount || 0) + 1;
-
-  dropdown.remove(); // Remove the dropdown menu
-  dropdown.style.display = "none";
-}
-
-function handleElseOption(block, elifElseDiv, dropdown, plusIcon) {
-  if (elifElseDiv.children.length === 1 && elifElseDiv.contains(plusIcon)) {
-    elifElseDiv.remove(); // Remove the plus icon container if it exists
-  }
-
-  // Create the new "else" block
-  const newElifElseDiv = createElifElseDiv("else");
-  const elseSpan = createSpan("else:");
-  newElifElseDiv.appendChild(elseSpan);
-
-  // Append the new "else" block to the parent block
-  block.appendChild(newElifElseDiv);
-  appendChildElement(block, "block");
-
-  // Move the plus icon to the end
-  const plusIconDiv = block.querySelector('.elif-else[data-elif-else-type="plus"]');
-  if (plusIconDiv) {
-    block.appendChild(plusIconDiv); // Move the plus icon to the end
-  } else {
-    // If the plus icon doesn't exist, create it and append it to the end
-    const newPlusIcon = createPlusIcon();
-    const newElifElseDivForPlus = createElifElseDiv("plus");
-    newElifElseDivForPlus.appendChild(newPlusIcon);
-    block.appendChild(newElifElseDivForPlus);
-
-    // Set up the dropdown menu for the new plus icon
-    setupDropdownMenu(newPlusIcon, block, newElifElseDivForPlus);
-  }
-
-  // Reset and update the IDs for all elif-else blocks
-  resetAndUpdateElifElseIds(block);
-
-  // Increment the "else" counter
-  block.dataset.elseCount = parseInt(block.dataset.elseCount || 0) + 1;
-
-  dropdown.remove(); // Remove the dropdown menu
-  dropdown.style.display = "none";
-}
-
-// ==========================
-// 5. Create Functions
-// ==========================
-
-function createInputField(placeholder, className, dataKey, blockID) {
-  const input = document.createElement("input");
-  input.type = "text";
-  input.placeholder = placeholder;
-  input.classList.add(className);
-
-  input.addEventListener("input", function () {
-    const block = input.closest(".box");
-    const value = input.value.trim();
-
-    // Validate input based on block type
-    if (blockID === "mathText" || blockID === "mathBlock") {
-      // Only allow numbers for Math blocks
-      if (/^-?\d*\.?\d*$/.test(value)) {
-        block.dataset[dataKey] = value;
-      } else {
-        input.value = block.dataset[dataKey] || "";
-      }
-    } else {
-      // Allow any input for non-Math blocks
-      block.dataset[dataKey] = value;
-    }
-  });
-
-  return input;
-}
-
-function createOperatorDropdown(blockID) {
-  const dropdown = document.createElement("select");
-  dropdown.classList.add("block-dropdown");
-
-  const operatorOptions = getBlockDropdownList(blockID); // Fetch operators based on blockID
-  operatorOptions.forEach((op) => {
-    const option = document.createElement("option");
-    option.value = op;
-    option.textContent = op;
-    dropdown.appendChild(option);
-  });
-
-  dropdown.value = "---"; // Default value
-  dropdown.addEventListener("change", function () {
-    const block = dropdown.closest(".box");
-    block.dataset.blockOperator = dropdown.value;
-  });
-
-  return dropdown;
-}
-
-function createInputBlock(block, placeholder, className, dataKey, blockID) {
-  const inputField = createInputField(placeholder, className, dataKey, blockID);
-  block.appendChild(inputField);
-}
-
-function createDropdownMenu() {
-  const dropdown = document.createElement("div");
-  dropdown.classList.add("dropdown-menu");
-  dropdown.style.display = "none";
-  return dropdown;
-}
-
-function createDropdownOption(text, onClickHandler) {
-  const option = document.createElement("div");
-  option.textContent = text;
-  option.classList.add("dropdown-item");
-  option.addEventListener("click", onClickHandler);
-  return option;
-}
-
-function createElifElseDiv(type) {
-  const div = document.createElement("div");
-  div.classList.add("elif-else");
-  div.setAttribute("data-elif-else-type", type);
-  return div;
-}
-
-function createSpan(text) {
-  const span = document.createElement("span");
-  span.textContent = text;
-  return span;
-}
-
-function createPlusIcon() {
-  const icon = document.createElement("i");
-  icon.classList.add("fa-solid", "fa-plus");
-  return icon;
-}
-
-// ==========================
-// 6. Child Block Functions
-// ==========================
-
-function appendChildElement(block, childElement) {
-  if (childElement === "block") {
-    const childBox = document.createElement("div");
-    childBox.classList.add("child-box-container");
-    childBox.dataset.parentID = block.id;
-    childBox.dataset.parentBlockID = block.dataset.blockID;
-
-
-    childBox.dataset.blockDepth = parseInt(block.dataset.blockDepth) + 1;
-
-    // Append the child-box-container to the appropriate parent
-    if (block.dataset.blockID === "if") {
-      const elifElseDiv = block.querySelector(".elif-else:last-child");
-      if (elifElseDiv) {
-        elifElseDiv.appendChild(childBox);
-      } else {
-        block.appendChild(childBox);
-      }
-    } else {
-      block.appendChild(childBox);
-    }
-
-    if (block.dataset.blockID === "if") {
-      const existingElifElseDiv = block.querySelector(".elif-else");
-      if (!existingElifElseDiv) {
-        const elifElseDiv = document.createElement("div");
-        elifElseDiv.classList.add("elif-else");
-        const plusIcon = document.createElement("i");
-        plusIcon.classList.add("fa-solid", "fa-plus");
-        elifElseDiv.appendChild(plusIcon);
-        block.appendChild(elifElseDiv);
-
-        // Set up the dropdown menu for the plus icon
-        setupDropdownMenu(plusIcon, block, elifElseDiv);
-      }
-    }
-  }
-}
-
-function createChildBoxHorizontal(parentID, parentBlockID) {
-  const childBox = document.createElement("div");
-  childBox.classList.add("child-box-container-horizontal");
-  childBox.dataset.parentID = parentID;
-  childBox.dataset.parentBlockID = parentBlockID;
-  return childBox;
-}
-
-
-// ==========================
-// 7. Update Functions
-// ==========================
-function updateVariableAttributes(block, selectedVariable) {
-  const existingAttributes = block.querySelectorAll(".variable-attribute");
-  existingAttributes.forEach((attr) => attr.remove());
-
-  updateVariableValueInBlock(block, selectedVariable);
-}
-
-function updateOperatorAttributes(block, selectedOperator) {
-  const operatorLabel = document.createElement("span");
-  operatorLabel.classList.add("operator-attribute");
-  operatorLabel.textContent = `Operator: ${selectedOperator}`;
-
-  const existingOperatorLabel = block.querySelector(".operator-attribute");
-  if (existingOperatorLabel) {
-    existingOperatorLabel.remove();
-  }
-
-  block.appendChild(operatorLabel);
-}
-
-// May not be needed anymore (look at calculateDepth() function)
-function updateDepth(block, targetBlock, depthChange) {
-  const currentDepth = parseInt(block.dataset.blockDepth) || 0;
-  const newDepth = currentDepth + depthChange;
-  block.dataset.blockDepth = newDepth; // Update depth attribute
-
-  // Update depth display in the block's label
-  const depthInfo = block.querySelector(".block-depth-info");
-  if (depthInfo) {
-    depthInfo.textContent = ` Depth: ${newDepth}`;
-  }
-}
-
-function resetAndUpdateElifElseIds(block) {
-  // Get all elif-else blocks within the parent block
-  const elifElseBlocks = block.querySelectorAll('.elif-else[data-elif-else-type="if"], .elif-else[data-elif-else-type="elif"], .elif-else[data-elif-else-type="else"]');
-
-  // Reset and update the IDs sequentially
-  elifElseBlocks.forEach((elifElseBlock, index) => {
-    elifElseBlock.dataset.ifElifElseId = index + 1; // Set the ID for elif-else elements
-  });
-
-}
-
-function updateUserVariableDropdowns() {
-  const dropdowns = document.querySelectorAll(
-    ".block-dropdown[data-type='variable']"
-  );
-  dropdowns.forEach((dropdown) => {
-    dropdown.innerHTML = ""; // Clear existing options
-    userVariables.forEach((varName) => {
-      const option = document.createElement("option");
-      option.value = varName;
-      option.textContent = varName;
-      dropdown.appendChild(option);
-    });
-  });
-}
-
-// ==========================
-// 8. UI and Interactivity
-// ==========================
-
-// Function to toggle between showing and hiding block categories
-function toggleCategory(categoryId) {
-  const allCategories = document.querySelectorAll(".category-blocks");
-  allCategories.forEach((category) => {
-    if (category.id === categoryId) {
-      category.classList.toggle("hidden"); // Toggle visibility of the clicked category
-    } else {
-      category.classList.add("hidden"); // Hide all other categories
-    }
-  });
-}
-
-// Function to add interactivity to blocks
-function addBlockInteractivity(block) {
-  block.draggable = true;
-  block.addEventListener("dragstart", dragStart);
-  block.addEventListener("dragover", dragOver);
-  block.addEventListener("drop", drop);
-  block.addEventListener("click", selectBlock);
-}
-
-// Event handler for starting a drag event on a block
-function dragStart(event) {
-  dragged = event.target.closest(".box");
-  if (dragged) {
-    event.dataTransfer.effectAllowed = "move"; // Allow movement
-    event.dataTransfer.setData("text/plain", dragged.id); // Store the block ID
-  }
-}
-
-// Event handler for dragging over a block or container
-function dragOver(event) {
-  event.preventDefault();
-  const targetBlock = event.target.closest(".box, .child-box-container, .child-box-container-horizontal");
-
-  if (!targetBlock || targetBlock === dragged) return;
-
-  // Check if the target container is inside the dragged block
-  if (dragged.contains(targetBlock)) {
-    return;
-  }
-
-  clearDropHighlights(); // Clear previous highlights
-
-  // Highlight the target container or block
-  if (targetBlock.classList.contains("child-box-container") || targetBlock.classList.contains("child-box-container-horizontal")) {
-    targetBlock.classList.add("highlight-inside");
-  } else if (targetBlock.classList.contains("box")) {
-    const rect = targetBlock.getBoundingClientRect();
-    const offsetY = event.clientY - rect.top;
-
-    if (offsetY < rect.height / 2) {
-      targetBlock.classList.add("drop-above");
-    } else {
-      // Check if the target block is the last block in the container
-      const container = targetBlock.parentElement;
-      const lastBlock = container.lastElementChild;
-
-      if (targetBlock === lastBlock) {
-        // If it's the last block, allow dropping below it
-        targetBlock.classList.add("drop-below");
-      } else {
-        targetBlock.classList.add("drop-below");
-      }
-    }
-  }
-}
-
-// Event handler for handling the drop action
-function drop(event) {
-  event.preventDefault();
-  if (!dragged) return;
-
-  const targetBlock = event.target.closest(".box, .child-box-container, .child-box-container-horizontal");
-
-  if (!targetBlock || targetBlock === dragged) return;
-
-  if (targetBlock.classList.contains("highlight-inside")) {
-    // Drop inside a container
-    targetBlock.appendChild(dragged);
-  } else if (targetBlock.classList.contains("drop-above")) {
-    // Drop above a block
-    targetBlock.parentNode.insertBefore(dragged, targetBlock);
-  } else if (targetBlock.classList.contains("drop-below")) {
-    // Drop below a block
-    const container = targetBlock.parentElement;
-    const lastBlock = container.lastElementChild;
-
-    if (targetBlock === lastBlock) {
-      // If the target is the last block, append the dragged block after it
-      container.appendChild(dragged);
-    } else {
-      // Otherwise, insert the dragged block after the target block
-      targetBlock.parentNode.insertBefore(dragged, targetBlock.nextSibling);
-    }
-  }
-
-  // Recalculate and update the block's depth
-  const newDepth = calculateDepth(dragged);
-  dragged.dataset.blockDepth = newDepth;
-
-  // Update the depth for all nested blocks
-  updateNestedDepths(dragged);
-
-  clearDropHighlights();
-  dragged = null;
-}
-
-// Function to clear all drop highlights
-function clearDropHighlights() {
-  document.querySelectorAll(".drop-above, .drop-below, .highlight-inside").forEach((block) => {
-    block.classList.remove("drop-above", "drop-below", "highlight-inside");
-  });
-}
-
-// Function to calculate the depth of a block
-function calculateDepth(block) {
-  let depth = 0;
-  let parent = block.closest(".child-box-container, .child-box-container-horizontal");
-
-  while (parent) {
-    depth++;
-    parent = parent.parentElement.closest(".child-box-container, .child-box-container-horizontal");
-  }
-
-  return depth;
-}
-
-// Function to update the depth of all nested blocks
-function updateNestedDepths(block) {
-  const newDepth = calculateDepth(block);
-  block.dataset.blockDepth = newDepth;
-
-  const childBlocks = block.querySelectorAll(".box");
-  childBlocks.forEach((childBlock) => {
-    updateNestedDepths(childBlock); // Recursively update depth for each child block
-  });
-}
-
-// Event handler for ending a drag event
-function dragEnd() {
-  clearDropHighlights(); // Clear all highlights
-  dragged = null; // Reset dragged block
-}
-
-//test function to show depth
-function showDepth(block) {
-  const depthLabel = block.querySelector(".block-depth-info");
-  if (!depthLabel) {
-    const label = document.createElement("span");
-    label.classList.add("block-depth-info");
-    block.appendChild(label);
-  }
-  block.querySelector(".block-depth-info").textContent = `Depth: ${block.dataset.blockDepth}`;
-}
-
-// Event handler for selecting a block when clicked
-function selectBlock(event) {
-  const targetBlock = event.target.closest(".box");
-
-  if (!targetBlock) return; // Exit if no block is found
-
-  // Clear previous selection if any
-  if (highlightedBlock) {
-    highlightedBlock.classList.remove("selected");
-  }
-
-  // Highlight the clicked block
-  highlightedBlock = targetBlock;
-  highlightedBlock.classList.add("selected");
-
-  event.stopPropagation(); // Prevent click event from propagating
-}
-
-function toggleDropdownVisibility(dropdown) {
-  dropdown.style.display = dropdown.style.display === "none" ? "block" : "none";
-}
-
-function setupDropdownMenu(plusIcon, block, elifElseDiv) {
-  const dropdown = createDropdownMenu();
-  block.appendChild(dropdown);
-
-  // Initialize counters for else if and else blocks using the block's dataset
-  if (!block.dataset.elseIfCount) block.dataset.elseIfCount = 0;
-  if (!block.dataset.elseCount) block.dataset.elseCount = 0;
-
-  // Function to update the dropdown options
-  const updateDropdownOptions = () => {
-    dropdown.innerHTML = "";
-
-    // Always add the "else if" option
-    const elseIfOption = createDropdownOption("else if", () =>
-      handleElseIfOption(block, elifElseDiv, dropdown, plusIcon)
-    );
-    dropdown.appendChild(elseIfOption);
-
-    // Add the "else" option only if there is no else block yet
-    if (block.dataset.elseCount === "0") {
-      const elseOption = createDropdownOption("else", () =>
-        handleElseOption(block, elifElseDiv, dropdown, plusIcon)
-      );
-      dropdown.appendChild(elseOption);
-    }
-  };
-
-  // Update dropdown options every time the dropdown is opened
-  plusIcon.addEventListener("click", () => {
-    updateDropdownOptions();
-    toggleDropdownVisibility(dropdown);
-  });
-}
 
 // ==========================
 // 9. Python Code Conversion
@@ -1230,60 +359,59 @@ function setupColumnResizing() {
 
 function setupDraggableBlocks() {
   document.querySelectorAll(".box").forEach((box) => {
-    box.draggable = true;
-    box.addEventListener("dragstart", dragStart);
-    box.addEventListener("dragover", dragOver);
-    box.addEventListener("drop", drop);
-    box.addEventListener("dragend", dragEnd);
+      box.draggable = true;
+      box.addEventListener("dragstart", dragStart);
+      box.addEventListener("dragover", dragOver);
+      box.addEventListener("drop", drop);
+      box.addEventListener("dragend", dragEnd);
   });
 
   // Prevent category buttons from closing the menu
   document.querySelectorAll(".category-blocks button").forEach((button) => {
-    button.addEventListener("click", (event) => {
-      event.stopPropagation();
-    });
+      button.addEventListener("click", (event) => {
+          event.stopPropagation();
+      });
   });
 
   // Handle block deletion when dragging over the left-side container
   const codeContainer = document.querySelector(".code-container");
   codeContainer.addEventListener("dragover", function (event) {
-    event.preventDefault();
+      event.preventDefault();
   });
 
   codeContainer.addEventListener("drop", function (event) {
-    event.preventDefault();
-    if (dragged) {
-      dragged.remove();
-      dragged = null;
-      updateLineNumbers();
-    }
+      event.preventDefault();
+      if (dragged) {
+          removeBlock(dragged.id); // Use removeBlock to delete the block
+          dragged = null;
+      }
   });
 
   // Remove drop target highlight when dragging leaves a block
   document.addEventListener("dragleave", function (event) {
-    const targetBlock = event.target.closest(".box");
-    if (targetBlock) {
-      targetBlock.classList.remove("drop-target");
-    }
+      const targetBlock = event.target.closest(".box");
+      if (targetBlock) {
+          targetBlock.classList.remove("drop-target");
+      }
   });
 
   // Deselect block when clicking outside
   document.addEventListener("click", function (event) {
-    if (highlightedBlock && !highlightedBlock.contains(event.target)) {
-      highlightedBlock.classList.remove("selected");
-      highlightedBlock = null;
-    }
+      if (highlightedBlock && !highlightedBlock.contains(event.target)) {
+          highlightedBlock.classList.remove("selected");
+          highlightedBlock = null;
+      }
   });
 
   // Delete highlighted block with "Delete" key
   document.addEventListener("keydown", function (event) {
-    if (event.key === "Delete" && highlightedBlock) {
-      highlightedBlock.remove();
-      highlightedBlock = null;
-      updateLineNumbers();
-    }
+      if (event.key === "Delete" && highlightedBlock) {
+          removeBlock(highlightedBlock.id); // Use removeBlock to delete the block
+          highlightedBlock = null;
+      }
   });
 }
+
 
 // ==========================
 // 13. Miscellaneous Code
